+++ watch_connectivity/ios/Classes/SwiftWatchConnectivityPlugin.swift
@@ -1,5 +1,4 @@
 import Flutter
-import HealthKit
 import UIKit
 import WatchConnectivity
 
@@ -52,16 +51,6 @@ public class SwiftWatchConnectivityPlugin: NSObject, FlutterPlugin, WCSessionDel
                 result(nil)
             } catch {
                 result(FlutterError(code: "Error updating application context", message: error.localizedDescription, details: nil))
-            }
-        case "startWatchApp":
-            HKHealthStore().startWatchApp(with: HKWorkoutConfiguration()) { success, error in
-                if success {
-                    result(nil)
-                } else if let error = error {
-                    result(FlutterError(code: "Error starting watch app", message: error.localizedDescription, details: nil))
-                } else {
-                    result(FlutterError(code: "Unable to start watch app", message: nil, details: nil))
-                }
             }
         
         // Not implemented