+++ watch_connectivity_platform_interface/lib/src/watch_connectivity_platform_interface_base.dart
@@ -90,9 +90,4 @@ abstract class WatchConnectivityBase {
   Future<void> updateApplicationContext(Map<String, dynamic> context) {
     return channel.invokeMethod('updateApplicationContext', context);
   }
-
-  /// Start the watch app
-  Future<void> startWatchApp() {
-    return channel.invokeMethod('startWatchApp');
-  }
 }